Combine alerts with modal mod
