+++ src/renderer/coremods/settings/pages/Addons.tsx
@@ -1,4 +1,4 @@
-import { api, confirm, fluxDispatcher, users } from "@common";
+import { api, fluxDispatcher, modal, users } from "@common";
 import React from "@common/react";
 import { Button, Divider, Flex, Input, SwitchItem, Text } from "@components";
 import { RepluggedPlugin, RepluggedTheme } from "src/types";
@@ -284,7 +284,7 @@ function Cards({
             setDisabled(clonedDisabled);
           }}
           uninstall={async () => {
-            const confirmation = await confirm.confirm({
+            const confirmation = await modal.confirm({
               title: `Uninstall ${addon.manifest.name}`,
               body: `Are you sure you want to uninstall this ${label(
                 type,
@@ -293,7 +293,6 @@ function Cards({
               cancelText: "Cancel",
               confirmColor: Button.Colors.RED,
             });
-            console.log(confirmation);
             if (!confirmation) return;
 
             const manager = getManager(type);
